backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.35.28
botocore==1.35.28
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
jmespath==1.0.1
multidict==6.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
tomli==2.1.0
twilio==9.3.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
channels_redis==4.1.0
urllib3==2.2.3
yarl==1.11.1
zope.interface==7.1.1
